+++ code/src/sixsq/nuvla/server/resources/module/utils.clj
@@ -1,13 +1,10 @@
 (ns sixsq.nuvla.server.resources.module.utils
   (:require
-    [clj-yaml.core :as yaml]
     [clojure.edn :as edn]
-    [clojure.set :as set]
     [clojure.string :as str]
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.pricing.impl :as pricing-impl]
     [sixsq.nuvla.server.resources.common.crud :as crud]
-    [sixsq.nuvla.server.util.log :as logu]
     [sixsq.nuvla.server.util.response :as r]))
 
 (def ^:const subtype-comp "component")
@@ -74,61 +71,6 @@
       (if (some? (nth versions i))
         i
         (recur (dec i))))))
-
-
-(def ^:const compose-specific-keys
-  #{"devices" "build" "cap_add" "cap_drop" "cgroup_parent" "container_name" "depends_on"
-    "external_links" "network_mode" "restart" "security_opt" "tmpfs" "userns_mode" "privileged"
-    "domainname" "ipc" "mac_address" "shm_size"})
-
-
-(def ^:const swarm-specific-keys #{"configs" "deploy"})
-
-
-(defn docker-compose-services-keys-set
-  [{:strs [services] :as _docker-compose}]
-  (set
-    (when (coll? services)
-      (->> services
-           (map (fn [[_ service-content]]
-                  (when (map? service-content)
-                    (-> service-content keys set))))
-           (reduce set/union)))))
-
-
-(defn list-swarm-unsupported-options
-  [services-keys-set]
-  (set/intersection services-keys-set compose-specific-keys))
-
-
-(defn some-services-has-swarm-options?
-  [services-keys-set]
-  (boolean (some swarm-specific-keys services-keys-set)))
-
-(defn parse-and-throw-when-not-parsable-docker-compose
-  [docker-compose]
-  (try
-    (yaml/parse-string docker-compose :keywords false)
-    (catch Exception _
-      (logu/log-and-throw-400 "Server wasn't able to parse docker-compose yaml!"))))
-
-
-(defn get-compatibility-fields
-  [docker-compose]
-  (let [services-keys-set   (docker-compose-services-keys-set docker-compose)
-        unsupported-options (-> services-keys-set list-swarm-unsupported-options vec)]
-    (if (and (not (some-services-has-swarm-options? services-keys-set))
-             (seq unsupported-options))
-      ["docker-compose" []]
-      ["swarm" unsupported-options])))
-
-
-(defn parse-get-compatibility-fields
-  [subtype docker-compose]
-  (when (is-application? subtype)
-    (-> docker-compose
-        parse-and-throw-when-not-parsable-docker-compose
-        get-compatibility-fields)))
 
 
 (defn retrieve-content-id