+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Changed
 
+- Module - Let user set compatibility flag
+- Module application - Deprecate unsupported-options
+- ES binding - Support bulk-edit
+- Spec helper - Generic spec validator
+- Spec helper - Request body validator
+- Common body - Spec for common request body specs
+- Nuvlabox - Adds bulk editing endpoints for tags
 - Module - Resolve vendor email
 
 ## [6.0.11] - 2023-02-22